Successful /discos/{uri}/latest request now returns 302 response
The /discos/{uri}/latest path will be Memento's URI-R and URI-G path.
To prepare for this, instead of automatically providing the RDF for the
latest version of the DiSCO in the response body with a 200 OK message,
this will be a 2-step process. The first response will be 302 Found with
the target DiSCO path shown in the Location header.  This makes it
consistent with how Memento works.