Added version date support, removed DiSCODTO logic
In preparation for Memento functionality, this change updates the
version list handling. There is now an option to retrieve a version list
with dates. The changes made DiSCODTO more cubersome than helpful, so
removed that functionality. Extracted link builder function to an inner
class, in the disco response manager. This is a just a first step to
cleaning this up... it needs more work especially if complex link
headers are also added to Resources path.  Wrote tests for these
changes, including some improvements to update tests to check
next/previous/last DiSCO headers.