Popup information in visualization; Shortened edge labels (#21)
Removed literals from Graph visualization, instead they are visible in a popup when you click on the node and are retrieved using a jquery GET. You can also use this popup to recenter the graph or navigate to the
resource's URL. This required corresponding work in core to support the retrieval of literals filtered by context (e.g. if you are looking at a DiSCO you need to limit the scope of the triples retrieved.) Note this
change also means that all literals are displayed when you click on a node in /resources/ view. Previously you would have needed to navigate to that node to see the literals associated.

Shortened edge labels to show no prefix.

Removed HTML from literals so that it doesn't interfere with display