Added external link icon and links to resource and discos view
Some links load RMap graphs, others load external pages. Have added
icons to improve this distinction.
In properties lists, made certain predicates bubble to the top if
present  (name, identifier, title, label). This is especially useful in
the node info popup as it brings key info to the top of the display.