Added support for customizable icons in visualization
Added an option to use an image instead of a dot or square in the
visualization.  Also created a set of icons that will be used by
default.

Note that the weight option on the node had to be disabled - it doesn't
seem to work proportionally on either the dots or the images - it is
either large or small. While this was OK on the dots, it made the icons
too small. May be worth revisiting in the future to get it working.