Moved pagination logic to core
Started to change core pagination logic to support GUI pagination work.
Pagination logic was in the API app, now instead of passing back a list
of URIs or triples from RMapService, where pagination might be
triggered, a ResultBatch is passed back. This contains information about
the position of the batch relative to the broader recordset and supports
pagination functions.

Moved popup nodeinfo js to the nodesedges.js.