Separated table data from graph data in GUI.  JSP cleanup. GUI pagination is a WIP
To support pagination, was necessary to separate table data from graph
data since they will paginate in different ways.  Added tabs for this to
Agent, Event, and DiSCO display. Table data is now loaded in the
background after the graph has loaded, but need more work to separate
out the model objects since the required data is different.
Added the option to exclude literals and/or types from resource queries
so that pagination count can still be maintained by core even when these
are removed.
Reduced size of page header so that more data visible when page first
loads.
Since the pages are getting more complicated, I created JSTL "tags" to
simplify the code and make link display more consistent.  Need to do
more of this, and split some of the left nav displays out into separate
pages that are loaded independently.