Update the SesameHttpTriplestore to use the TriplestoreInitializer from the 'spring-util' module.
If a TriplestoreInitializer instance has been set, it will be invoked inside 'initializeRepository()'.  The default implementation of TriplestoreInitializer will attempt to create an instance of a Sesame HTTP triplestore by invoking the OpenRDF Workbench webapp.

Ideally, triplestore initialization could be handled at application startup (rather than "just in time"), but the main challenge to startup initialization is that the API to Sesame is over HTTP.  If the RMap webapplications and Sesame are managed by the same servlet container instance, the RMap webapp cannot call over to the Sesame  webapp during initialization due to the Servlet spec and lifecycle concerns.  So for now, this JIT approach will have to do.  Since the SesameHttpTriplestore class already has the notion of just-in-time initialization (expressed by 'initializeRepository()'), it wasn't a stretch to include the TriplestoreInitializer here as well.