Configure integration environment and update ITs (#79)
This is a large PR which updates RMap to support an integration environment, complete with a HTTP-based Sesame triplestore and a Derby database which is shared between the RMap API and HTML UI webapps.  This allows for complete integration testing between the RMap webapps.

The `integration` module is configured to start a Derby database, then start Tomcat.  Tomcat starts the two RMap webapps, as well as the OpenRDF Sesame HTTP server and Workbench.

`mvn verify` can be used in the `integration` module to run the ITs.

`mvn validate cargo:run` can be used in the `integration` module to start an interactive developer runtime.

`mvn clean` will remove any data left over from the ITs or from interacting with the developer runtime.  Otherwise, data is persisted between invocations of the ITs or developer runtime.

Two new Spring profiles are added: `integration-db` and `integration-triplestore`.  The integration profiles are wired such that they are compelled to preserve existing data, otherwise they initialize their respective data stores.  These integration profiles reside in the `integration` module, but use Spring `<import...` statements from the `auth` and `core` modules' Spring contexts.

Updated developer documentation is found in `DEVELOPER.md`.