Added tests for ORMapEventMgr, fixed issue where API Key wasn't being included in Event
Added a number of unit tests for ORMapEventMgr, which previously had
none.
Fixed issue #89 - there is an option to include the API Key in the
Event, this was getting added to the graph, but not getting read back.
This is fixed, and there are tests for it.
While writing tests, found another issue where
ORMapEventUpdate/ORMapEventDerviation's createdObjectIds weren't being
populated in some situations. I think this might be the cause of issue
#96, but will need further tests to confirm.