Can now disable some or all oauth options using new properties
- Added properties rmapweb.google-oauth-enabled,
rmapweb.twitter-oauth-enabled, and rmapweb.orcid-oauth-enabled. These
default to false and the system will check them in order to determine
whether the corresponding OAuth buttons will appear. Where all 3 OAuth
options are disabled the login link will disappear completely. Any
attempt to reach a login page will redirect to the home page.  The
properties that determine whether each oauth option is enabled affects
the layout of the view so they are pulled in using a new SiteProperties
class, and this object is added to the model on each page load using an
HandlerInterceptor.postHandle().  This can be used for other site-wide
properties in the future.
- Added some test for new interceptor
- Added simple test for HomeController.