GUI changes for a User Admin tool to manage users and keys
Series of changes for issue #109 (Add a User Admin tool to RMap)
- Added an admin login page, whose user/pass is set in the
rmap.properties.  Logging in as admin kicks out any rmap user logged in
to that session.   Added admin menu.
- Added users list that can be paginated through and filtered
- Added support to create or edit a user
- Added support to create and edit user keys
- Separated out /admin/ path from oauth user path.  The admin tool uses
many of the same views as the oauth options does, but this change ensure
all admin pages are accessed through the {rmap-gui-url}/admin/* path
thus supporting additional options to configure access to the tool based
on path.