User admin GUI, Agent creation work on API app, other refactoring
- Completed user admin GUI work - can create users and keys
- Did corresponding work on API - refactored how Agents are initiated in
the RMap database to correlate with possibility of admin user creation.
Previously AgentIds were being minted when the user was set to sync with
RMap, but this could be wasteful of IDs, changed it to mint when first
create a DiSCO.
- Rearranged UserRMapAgentServiceImpl to ApiUserServiceImpl to require
less dependencies and to split out some functionality that was lumped
together.
- Fixed and rewrote tests for ApiUserServiceImpl to use Mockito instead
of an altered copy of the class. This eliminated the need for the
persistent-userservice/mock-userservice profiles.
Note: some integration tests not currently work, still need work here.