GUI changes for RMap admin tool, changes to Agent creation, tests for API key (#122)
* GUI changes for a User Admin tool to manage users and keys

Series of changes for issue #109 (Add a User Admin tool to RMap)
- Added an admin login page, whose user/pass is set in the
rmap.properties.  Logging in as admin kicks out any rmap user logged in
to that session.   Added admin menu.
- Added users list that can be paginated through and filtered
- Added support to create or edit a user
- Added support to create and edit user keys
- Separated out /admin/ path from oauth user path.  The admin tool uses
many of the same views as the oauth options does, but this change ensure
all admin pages are accessed through the {rmap-gui-url}/admin/* path
thus supporting additional options to configure access to the tool based
on path.

* User admin GUI, Agent creation work on API app, other refactoring

- Completed user admin GUI work - can create users and keys
- Did corresponding work on API - refactored how Agents are initiated in
the RMap database to correlate with possibility of admin user creation.
Previously AgentIds were being minted when the user was set to sync with
RMap, but this could be wasteful of IDs, changed it to mint when first
create a DiSCO.
- Rearranged UserRMapAgentServiceImpl to ApiUserServiceImpl to require
less dependencies and to split out some functionality that was lumped
together.
- Fixed and rewrote tests for ApiUserServiceImpl to use Mockito instead
of an altered copy of the class. This eliminated the need for the
persistent-userservice/mock-userservice profiles.
Note: some integration tests not currently work, still need work here.

* Undo UserRMapAgentServiceImpl constructor change

Having problems with integration tests during Travis build and different
problems building locally. Checking if this reversal fixes the problem
so I can hone in what's wrong.

* Change User record to eager load idProviders and Keys

Integration test was failing due to there no longer being a connection
to do a lazy load for User record id providers and keys. Changed to
eager load since these lists will typically be short and to avoid
unexpected behavior when passing around User object

* Added missing page in AdminToolController

* Fixed comment.

Said "HomeController" instead of AdminToolController.

* Fixed comment

* Fixed javadoc

* Minor code changes for PR #122 review feedback

- Added hashCode() override to AdminLogin
- Made some fields private
- Set logback-test.xml back to WARN level
- Used @Test(expected=RMapObjectNotFoundException.class) annotation for
test with nothing to do in catch

* Changed default properties for admin tool

- changed admin tool to default to disabled
- added admin tool properties to integration test (defaults to enabled)

* Made admin-tool-enabled property work

- rmapweb.admin-tool-enabled was not wired up to work - this change
fixes it.
- A few minor javadoc changes
- merged /admin/user/settings into /user/settings
- fixed handling of api key page for admin use