Resolve connection hang in webapp (#123)
* Resolve connection hang in webapp

- Connection was hanging due to apparent quirk with retrieving
statements that contain a Resource that doesn't exist in the database.
This leaves the connection open and uses up the concurrent connections
allowed.  This was specifically a problem when checking for an RMap type
using ORMapObjectMgr.isRMapType().  Now checks the context exists before
checking the type statement.
- Some logging added.
- Added code to ensure all connections close even when code errors
- Change ORMapService to use only autowired constructor

* Move connectionClose responsibility to ORMapService

Per discussion PR #123, this change moves the finally
{closeConnection()} logic to ORMapService and removes it from webapp,
api and auth.  The apps over core should no longer concern themselves
with connection closes.

* Fixes for travis build - one more closeConnection() removed!