Change how user record is loaded to deal with error resulting from log change
Removing a "toString()" from one of the log messages revealed this
problem.  The toString was actually triggering an object not found
exception for the User record that was the making the system function as
expected! The parameterized log message doesn't trigger this and caused
tests to fail.  Modified to use session.get() and check for null.
Looked over all calls to this method to ensure it would still work as
expected.