Added hard delete DiSCO function to rmap-core
Previously there was no way to delete a DiSCO, only tombstone it.
Delete removes the DiSCO statements from the triplestore.  As part of
this change:
- Changed delete event to only support one deleted object, same as
tombstone
- Added a property to define an administrator agent uri. Either this
Agent or the original creating Agent can perform deletes and tombstones
of DiSCOs. Note that in the API we chose not to allow hard deletes, but
this is now possible.  The hard delete function will be performed by an
administrator in the webapp admin tool.
- Added some tests for this change