Rename RMapRequestAgent to RequestEventDetails, expand to include description (#128)
RMapRequestAgent is supplied to rmap-core when triplestore changes are made. It previously only contained systemAgentUri and keyUri. Both of these fields end up in the RMap Event corresponding to the change. There wasn't previously a way for client apps to provide Event descriptions when they perform updates. This change expands RMapRequestAgent to include all client provided Event information (i.e. it adds the description field). This allows client apps to put a description in the Event description field. This is being added to support work on hard-deletes through the admin tool since the GUI will allow admins to enter a note about the hard delete.