Add tooltips and glossary; change node type toggle to checkbox.
These changes respond to the issue of users not understanding what a
DiSCO, Agent etc. is. A common behavior in usability testing was to look
for definitions, or go to the legend in the visualization for
definitions. In the case of the visualization, clicking for a definition
would unexpectedly remove part of the graph.  The changes attempt to
deal with these problems.
- Added glossary of common terms in RMap
- Added tooltips next to terms that may be ambiguous or confusing
- Added tooltips to the legend
- Converted the toggle type functionality to checkboxes in the legend
- Changed Uncategorized type to Other
- Removed uneccessary types "Literal" and "Type"