Deal with problem of lazy load in some database connections
Investigated this after noticing that when using in-memory derby, an
error occurs on the key download page saying can't lazy load the data.
It does not occur on live database tests.  To resolve this, made the
following changes:

- Force lazy=false on the database model. My best guess at what is
happening is that for some reason different connections have different
default behaviors wrt ending session on commit. In particular, it seems
the in-memory derbydb ends the session on commit which causes anything
that has lazy load not to load when retrieved after commit e.g.
ApiKey.getUserId. This happens in the ApiKeyController.downloadKey
method.  We can either disable session close on commit for each
connection, or just force lazy=false by adding an annotation to the data
model classes - this makes it OK to close session when you're done with
the database.  Rather than have to keep dealing with different behaviors
via spring, I opted to add lazy=false since the data model is small and
any eager load will not have to load much data, even where there are
joins.
- Added a test to ensure downloadKey working
- Added a null check on site properties