Fixed validation of session userId against key userId
In ApiKeyController.updateUserKey() does a check to make sure the
session user matches the user of the key being updated. It was failing
because it was looking at form data for user id, but it needed to get
the full record for apiKey from the database.  Loaded record from
database to use in validation.