Updated Sesame OpenRDF to Eclipse RDF4J (#182)
This closes issue #71.

Last year, OpenRDF / Sesame Framework became Eclipse RDF4J. This change switches RMap to use the most recent Eclipse RDF4J (2.2.2). Some paths had to be renamed from openrdf to rdf4j. Several property names have changed as well.

The only remnant from openrdf is that many classes are still prefixed with "O" representing the OpenRdf implementation. A little odd, but this would require a change to about 25 more file names. I have made this a separate issue #184 to be addressed once other PRs are merged and we are closer to a release.