Implement a Kafka Partitioner that insures events pertaining to a lineage go to the same Kafka partition.
Note that this implementation is sensitive to the number of Kafka partitions that are configured for the RMap event topic.  If the number of partitions changes, then the algorithm implemented in this commit will end up assigning events that once went to PartitionA to go to PartitionB.  This will break the guarantee that events pertaining to a lineage are contained within a single partition.  Care should be taken if the number of partitions for the RMap event topic are ever increased in the future.

This closes #144 and #138.