Assign RMap events to Kafka partitions based on progenitor lineage URI (#195)
This implements a Kafka Partitioner that insures events pertaining to a lineage go to the same Kafka partition.  Events that do not have a progenitor lineage URI are simply assigned based on the hash code of the event object.

Note that this implementation is sensitive to the number of Kafka partitions that are configured for the RMap event topic.  If the number of partitions changes, then the algorithm implemented in this commit will end up assigning events that once went to PartitionA to go to PartitionB.  This will break the guarantee that events pertaining to a lineage are contained within a single partition.  Care should be taken if the number of partitions for the RMap event topic are ever increased in the future.

Includes unit test.

This closes #144 and #138.