Removed double HTTP decoding of parameters passed into API or GUI (#240)
* Path and PathVariable annotations automatically decode, removed double decode

Closes #230
When a Resource URI has special characters encoded into the querystring,
they were being double decoded in RMap and this was causing matches not
to be found. The decoding already happens by default when the Path
annotations are used. This removes the additional http decoding which
was causing problems when using the encoded parameter URLs.
It also replaces escaped quotes with regular quotes in results that are
pulled from Solr when searching for labels.

* Updated tests so that they don't use encoded URIs
Fixed minor alignment issue on searchpage.